% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixpanel.R
\name{mp_default_props}
\alias{mp_default_props}
\title{Set default properties for Mixpanel events}
\usage{
mp_default_props(properties)
}
\arguments{
\item{properties}{List of properties.}
}
\description{
These properties will be sent with every subsequent Mixpanel event. Calling this function
multiple times will override the previous defaults.
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinymixpanel)

  ui <- fluidPage(
    mp_init(YOUR_PROJECT_TOKEN)
  )

  server <- function(input, output, session) {
    mp_default_props(list(foo = "bar", shiny_version = as.character(packageVersion("shiny"))))
    mp_track("page init")
  }

  shinyApp(ui, server)
}
}
