% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixpanel.R
\name{mp_default_props_js}
\alias{mp_default_props_js}
\title{Set default client-side properties for Mixpanel events}
\usage{
mp_default_props_js(properties)
}
\arguments{
\item{properties}{List of properties. The value of each element in the list will be
treated as JavaScript code to be evaluated in the client's browser.}
}
\description{
These properties will be sent with every subsequent Mixpanel event. The properties
will be computed client-side, on the user's browser.
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinymixpanel)

  ui <- fluidPage(
    mp_init(YOUR_PROJECT_TOKEN)
  )

  server <- function(input, output, session) {
    mp_default_props_js(list(size = "screen.width", ua = "navigator.userAgent"))
    mp_track("page init")
  }

  shinyApp(ui, server)
}
}
