% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_track_server.R
\name{mp_track_server}
\alias{mp_track_server}
\title{Track an event to Mixpanel using the server-side API}
\usage{
mp_track_server(event, properties = list(), userid = "", token = "")
}
\arguments{
\item{event}{Name of the event.}

\item{properties}{List of properties to send for this event.}

\item{userid}{A user ID to identify with this event.}

\item{token}{The Mixpanel project token. If not provided, then an environment variable
\code{SHINYMIXPANEL_TOKEN} will be used.}
}
\description{
Using the client-side API (\verb{[mp_track()]}) is preferable because sending data through the web
browser automatically includes more data about the user. The advantage of using the server-side
API is that it cannot be blocked with an ad blocker.\cr\cr
The server-side API can also be used by any R script outside of a Shiny app, while the client-side
API can only be used in a Shiny app.
}
\examples{
\dontrun{
  Sys.setenv("SHINYMIXPANEL_TOKEN" = YOUR_PROJECT_TOKEN)
  mp_userid("abcd1234")
  mp_default_props(list("foo" = "bar", "text" = "hello"))
  mp_track_server("greet", list(name = "dean"))
}
}
