% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixpanel.R
\name{mp_userid}
\alias{mp_userid}
\title{Associate all Mixpanel events to a user}
\usage{
mp_userid(userid)
}
\arguments{
\item{userid}{A user ID to identify with Mixpanel.}
}
\description{
After calling \code{mp_userid()}, all subsequent Mixpanel events will be associated with
this user ID.
}
