% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixpanel.R
\name{mp_init}
\alias{mp_init}
\title{Set up a Shiny app to use Mixpanel for event tracking}
\usage{
mp_init(
  token,
  userid = "",
  options = list(),
  default_properties = list(),
  default_properties_js = list(),
  test_token = "",
  test_domains = list("127.0.0.1", "localhost"),
  force = FALSE
)
}
\arguments{
\item{token}{The Mixpanel project token. If not provided, then an environment variable
\code{SHINYMIXPANEL_TOKEN} will be used.}

\item{userid}{A user ID to identify with Mixpanel on the current page. If provided, this user ID
will be associated with all event tracking calls. If the user ID is not known in the UI and
is only known in the server, call \verb{[mp_userid()]} from the server.}

\item{options}{List of configuration options to pass to Mixpanel. A full list of supported
options is available in \href{https://developer.mixpanel.com/docs/javascript-full-api-reference#mixpanelset_config}{Mixpanel's documentation}.
See the examples below for sample usage.}

\item{default_properties}{List of properties to send with every event.}

\item{default_properties_js}{List of properties that need to be computed client-side (with JavaScript
in the user's browser) to send with every event. See the examples below for sample usage.}

\item{test_token}{A Mixpanel project token for a test project. Must be used together with \code{test_domains}.
See the section below on "Using a test Mixpanel project for testing/development".}

\item{test_domains}{List of domains where the \code{test_token} should be used. Must be used together
with \code{test_token}. See the section below on "Using a test Mixpanel project for testing/development".}

\item{force}{By default, {shinymixpanel} attempts to send all event tracking via the client-side API.
This is the preferred way to use Mixpanel, as it automatically gathers some user data such as their
physical location and browser information. However, some users may have disabled Mixpanel in their browser
(for example using an AdBlock), and for these users it's not possible to connect to Mixpanel through
the browser. If you set \code{force = TRUE}, then {shinymixpanel} will attempt to send events to Mixpanel
using REST API calls when the client side is blocked.}
}
\description{
Call this function in a Shiny app's UI in order to initialize Mixpanel on the page. A project
token is required, which can either be provided using the \code{token} parameter or with the
\code{SHINYMIXPANEL_TOKEN} environment variable.
}
\note{
Tracking can be temporarily disabled, for example while developing or testing, by setting
the \code{SHINYMIXPANEL_DISABLE} environment variable to \code{"1"}.
}
\section{Using a test Mixpanel project for testing/development}{

While developing or testing your Shiny app, you can use the \code{SHINYMIXPANEL_DISABLE} envvar
to disable Mixpanel tracking. However, sometimes you may want to still have tracking but send the
data to a different "test" project rather than the real production Mixpanel project. The {shinymixpanel}
package supports this usecase via the \code{test_token} and \code{test_domains} parameters.\cr\cr
If both of these parameters are provided, then if the Shiny app is in a domain listed in the
\code{test_domains} list, the data will be sent to the project specified by \code{test_token} instead.
Note that the domains in \code{test_domains} are assumed to be suffixes. This means that if you provide
\code{"example.com"} as a test domain, then any user on \code{example.com} or \code{test.example.com} will use
the test project.
}

\examples{
if (interactive()) {

  ### The simplest way to initialize Mixpanel is simply by calling `mp_init(token)`
  library(shiny)
  library(shinymixpanel)
  ui <- fluidPage(
    mp_init(YOUR_PROJECT_TOKEN)
  )

  server <- function(input, output, session) {
    mp_track("page init")
  }

  shinyApp(ui, server)

  ### The exapmle below shows how to use the many different parameters

  library(shiny)
  library(shinymixpanel)

  ui <- fluidPage(
    mp_init(
      YOUR_PROJECT_TOKEN,
      userid = "d24ba28c3",
      options = list(debug = TRUE),
      default_properties = list(foo = "bar", shiny_version = as.character(packageVersion("shiny"))),
      default_properties_js = list(size = "screen.width", ua = "navigator.userAgent"),
      test_token = TOKEN_FOR_TEST_PROJECT,
      test_domains = list("127.0.0.1", "internal.mycompany.com")
    )
  )

  server <- function(input, output, session) {
    mp_track("page init")
  }

  shinyApp(ui, server)
}
}
