% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixpanel.R
\name{mp_track}
\alias{mp_track}
\title{Track an event to Mixpanel}
\usage{
mp_track(event, properties = list(), properties_js = list())
}
\arguments{
\item{event}{Name of the event}

\item{properties}{List of properties to send for this event}

\item{properties_js}{List of properties to be computed client-side (with JavaScript
in the user's browser)}
}
\description{
Track an event to Mixpanel
}
