% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixpanel.R
\name{mp_track}
\alias{mp_track}
\title{Track an event to Mixpanel}
\usage{
mp_track(event, properties = list())
}
\arguments{
\item{event}{Name of the event}

\item{properties}{List of properties to send for this event}
}
\description{
When called from a Shiny app, the Mixpanel client-side API is used. This is the preferred way
to use Mixpanel, as it automatically gathers some user data such as their physical location
and browser information. If you prefer to use the server-side API explicitly, use
\verb{[mp_track_server()]}.\cr\cr
When called outside a Shiny app, \verb{[mp_track_server()]} is used.
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinymixpanel)

  ui <- fluidPage(
    mp_init(YOUR_PROJECT_TOKEN)
  )

  server <- function(input, output, session) {
    mp_track("page init")
  }

  shinyApp(ui, server)
}
}
