% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_userid.R
\name{mp_userid}
\alias{mp_userid}
\title{Associate Mixpanel events with a specific user}
\usage{
mp_userid(userid)
}
\arguments{
\item{userid}{A user ID to identify with Mixpanel.}
}
\description{
After calling \code{mp_userid()}, all subsequent Mixpanel events will be associated with
this user ID.
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinymixpanel)

  ui <- fluidPage(
    mp_init(YOUR_PROJECT_TOKEN)
  )

  server <- function(input, output, session) {
    mp_userid("fa8762b3a")
    mp_track("page init")
  }

  shinyApp(ui, server)
}
}
